+++ types/ember-cli-mirage/index.d.ts
@@ -1,4 +1,7 @@
+/* eslint-disable no-use-before-define */
+import MirageModelRegistry from 'ember-cli-mirage/types/registries/model';
 import DS from 'ember-data';
+import EmberDataModelRegistry from 'ember-data/types/registries/model';
 import { Document } from 'osf-api';
 
 export { default as faker } from 'faker';
@@ -32,7 +35,13 @@ export interface Database {
 }
 
 export type Model<T> = {
-    [P in keyof T]: T[P] & { models: any[] };
+    [P in keyof T]: T[P] extends DS.Model & DS.PromiseObject<infer M> ? ModelInstance<M> :
+        T[P] extends DS.Model ? ModelInstance<T[P]> :
+        T[P] extends DS.PromiseManyArray<infer M> ? Collection<M> :
+        T[P] extends DS.Model[] & DS.PromiseManyArray<infer M> ? Collection<M> :
+        T[P] extends DS.Model[] ? Collection<T[P]> :
+        T[P] extends Date ? Date | string :
+        T[P];
 };
 
 interface ModelInstanceShared<T> {
@@ -52,9 +61,11 @@ interface ModelInstanceShared<T> {
 
 export type ModelInstance<T = AnyAttrs> = ModelInstanceShared<T> & Model<T>;
 
-interface Collection<T> {
+export interface Collection<T> {
     models: Array<ModelInstance<T>>;
+    length: number;
     modelName: string;
+    firstObject: ModelInstance<T>;
     update<K extends keyof T>(key: K, val: T[K]): void;
     save(): void;
     reload(): void;
@@ -64,14 +75,14 @@ interface Collection<T> {
 }
 
 interface ModelClass<T = AnyAttrs> {
-    new(attrs: T): ModelInstance<T>;
-    create(attrs: T): ModelInstance<T>;
-    update(attrs: T): ModelInstance<T>;
-    all(): Collection<T>;
-    find<S extends ID | ID[]>(ids: S): S extends ID ? ModelInstance<T> : Collection<T>;
-    findBy(query: T): ModelInstance<T>;
-    first(): ModelInstance<T>;
-    where(query: T | ((r: ModelInstance<T>) => boolean)): Collection<T>;
+    new<M = T>(attrs: Partial<ModelAttrs<M>>): ModelInstance<M>;
+    create<M = T>(attrs: Partial<ModelAttrs<M>>): ModelInstance<M>;
+    update<M = T>(attrs: Partial<ModelAttrs<M>>): ModelInstance<M>;
+    all<M = T>(): Collection<M>;
+    find<M = T, S extends ID | ID[] = ID>(ids: S): S extends ID ? ModelInstance<M> : Collection<M>;
+    findBy<M = T>(query: Partial<ModelAttrs<M>>): ModelInstance<M>;
+    first<M = T>(): ModelInstance<M>;
+    where<M = T>(query: Partial<ModelAttrs<M>> | ((r: ModelInstance<M>) => boolean)): Collection<M>;
 }
 
 export interface Schema {
@@ -143,8 +154,15 @@ export type resourceAction = 'index' | 'show' | 'create' | 'update' | 'delete';
 
 export type ModelAttrs<T> = {
     [P in keyof T]: T[P] extends DS.Model & DS.PromiseObject<infer M> ? ModelInstance<M> :
-        T[P] extends DS.Model ? ModelInstance<T[P]> : T[P];
+        T[P] extends DS.Model ? ModelInstance<T[P]> :
+        T[P] extends DS.PromiseManyArray<infer M> ? Array<ModelInstance<M>> :
+        T[P] extends DS.Model[] & DS.PromiseManyArray<infer M> ? Array<ModelInstance<M>> :
+        T[P] extends DS.Model[] ? Array<ModelInstance<T[P]>> :
+        T[P] extends Date ? Date | string :
+        T[P];
 };
+
+export type ModelRegistry = EmberDataModelRegistry & MirageModelRegistry;
 
 export interface Server {
     schema: Schema;
@@ -173,33 +191,33 @@ export interface Server {
     // passthrough(...paths: string[], verbs?: Verb[]): void;
     passthrough(...args: any[]): void;
 
-    create<T extends AnyAttrs = AnyAttrs>(
-        modelName: string,
+    create<T extends keyof ModelRegistry>(
+        modelName: T,
         ...traits: string[]
-    ): ModelInstance<T>;
-    create<T extends AnyAttrs = AnyAttrs>(
-        modelName: string,
-        attrs?: Partial<ModelAttrs<T>>,
+    ): ModelInstance<ModelRegistry[T]>;
+    create<T extends keyof ModelRegistry>(
+        modelName: T,
+        attrs?: Partial<ModelAttrs<ModelRegistry[T]>>,
         ...traits: string[]
-    ): ModelInstance<T>;
+    ): ModelInstance<ModelRegistry[T]>;
 
-    createList<T extends AnyAttrs = AnyAttrs>(
-        modelName: string,
+    createList<T extends keyof ModelRegistry>(
+        modelName: T,
         amount: number,
         ...traits: string[]
-    ): Array<ModelInstance<T>>;
-    createList<T extends AnyAttrs = AnyAttrs>(
-        modelName: string,
+    ): Array<ModelInstance<ModelRegistry[T]>>;
+    createList<T extends keyof ModelRegistry>(
+        modelName: T,
         amount: number,
-        attrs?: Partial<ModelAttrs<T>>,
+        attrs?: Partial<ModelAttrs<ModelRegistry[T]>>,
         ...traits: string[]
-    ): Array<ModelInstance<T>>;
+    ): Array<ModelInstance<ModelRegistry[T]>>;
 
     shutdown(): void;
 }
 
-export type TraitOptions = AnyAttrs & {
-    afterCreate?: (obj: ModelInstance<AnyAttrs>, svr: Server) => void,
+export type TraitOptions<M> = AnyAttrs & {
+    afterCreate?: (obj: ModelInstance<M>, svr: Server) => void,
 };
 
 export interface Trait<O extends TraitOptions = {}> {
@@ -207,7 +225,10 @@ export interface Trait<O extends TraitOptions = {}> {
     __isTrait__: true;
 }
 
-export function trait<O extends TraitOptions>(options: O): Trait<O>;
+export function trait<
+    M extends ModelRegistry[keyof ModelRegistry],
+    O extends TraitOptions<M> = TraitOptions<M>
+>(options: O): Trait<O>;
 
 // TODO when https://github.com/Microsoft/TypeScript/issues/1360
 // function association(...traits: string[], overrides?: { [key: string]: any }): any;
@@ -217,7 +238,7 @@ export function association(...args: any[]): any;
 export type FactoryAttrs<T> = {
     [P in keyof T]?: T[P] | ((index: number) => T[P]);
 } & {
-    afterCreate?(newObj: any, server: Server): void;
+    afterCreate?(newObj: ModelInstance<T>, server: Server): void;
 };
 
 export class FactoryClass {
@@ -238,3 +259,4 @@ export class JSONAPISerializer {
     serialize(object: ModelInstance, request: Request): SingleResourceDocument;
     normalize(json: any): any;
 }
+/* eslint-enable no-use-before-define */