+++ CHANGELOG.md
@@ -32,6 +32,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Mirage:
     - Slim down default scenario
     - Allow different set of scenarios to run based on local settings with `MIRAGE_SCENARIOS`
+    - types:
+        - `server.create(modelName, ...)` now looks up the `modelName` in the `ModelRegistry` and properly types the
+          return values as `ModelInstance<ModelFromTheRegistry>` and type checks any model properties passed in.
+        - same as above, but for `server.createList`
+        - traits now take a type argument (the model they are a trait for) which results in proper typing for
+          `afterCreate(model, server)` without requiring manual typing of its args.
+        - the `afterCreate` method of mirage factories is typed similarly to trait's `afterCreate`
+        - mirage model class methods now take a type argument and their return value will be properly typed.
+          (e.g `server.schema.users.find<User>()` will return `ModelInstance<User>`)
 - Services
     - `analytics` - allow toast-on-click to be used in production builds (when enabled in dev banner)
 - Components